requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
cloudinary
pdf2docx
PyPDF2
python-docx
transformers
torch
black
flake8